fix(server): metadata detect local and release verify build
