+++ misc/tests/build.sh
@@ -3,11 +3,15 @@ set -euo pipefail
 SETUP_CONFIG="${SETUP_CONFIG:-misc/tests/config.sh}"
 source "${SETUP_CONFIG}"
 
-echo ">>> Building operator image (${IMG})..."
-make docker-build IMG="${IMG}"
+if [ "${BUILD_OPERATOR:-true}" != "false" ]; then
+    echo ">>> Building operator image (${IMG})..."
+    make docker-build IMG="${IMG}"
+fi
 
-echo ">>> Building gateway image (${SERVER_IMG})..."
-make docker-build-server SERVER_IMG="${SERVER_IMG}"
+if [ "${BUILD_SERVER:-true}" != "false" ]; then
+    echo ">>> Building server image (${SERVER_IMG})..."
+    make docker-build-server SERVER_IMG="${SERVER_IMG}"
+fi
 
 echo ">>> Loading images into kind..."
 kind load docker-image "${IMG}" --name "${CLUSTER_NAME}"
